4 bug fixed

+++ controllers/index.controller.js
@@ -9,6 +9,20 @@ exports.hello = async function (req, res) {
     try {
         const helper = {}
         const jsonObj = await csv({ checkType: true }).fromFile(req.file.path);
+        jsonObj.forEach(item => {
+            if (!item["Product ID"]) {
+                res.render('sessions', { covertingFailed: 'Converting Failed' })
+            }
+            if (!item["Product Name"]) {
+                res.render('sessions', { covertingFailed: 'Converting Failed' })
+            }
+            if (!item.Price) {
+                res.render('sessions', { covertingFailed: 'Converting Failed' })
+            }
+            if (!item.Quantity) {
+                res.render('sessions', { covertingFailed: 'Converting Failed' })
+            }
+        })
         const result = jsonObj.reduce((accumulator, currentValue) => {
             var key = currentValue['Product ID']
             if (!helper[key]) {
@@ -23,27 +37,37 @@ exports.hello = async function (req, res) {
             ...item,
             'Total Price': item['Price'] * item['Quantity'],
         }))
-        const dataObj = {
-            "_id": new mongoose.Types.ObjectId(),
-            "data": totalproductvalue
-        }
-        const newData = new Data(dataObj)
-        newData.save((err, data) => {
-            if (err) {
-                res.render('sessions', { failed: 'File Uploaded Failed' })
-                fs.unlinkSync(req.file.path)
-            } else {
-                const sessObj = {
-                    "_id": new mongoose.Types.ObjectId(),
-                    "created_by": req.session.passport.user,
-                    "data": data._id
-                }
-                const newSess = new Sessions(sessObj)
-                newSess.save()
-                res.render('sessions', { succes: 'Session Created Succesfully' })
-                fs.unlinkSync(req.file.path)
+        if (totalproductvalue.length === 0) {
+            res.render('sessions', { failed: 'File Uploaded Failed' })
+        } else {
+            const dataObj = {
+                "_id": new mongoose.Types.ObjectId(),
+                "data": totalproductvalue
             }
-        })
+            const newData = new Data(dataObj)
+            newData.save((err, data) => {
+                if (err) {
+                    res.render('sessions', { failed: 'File Uploaded Failed' })
+                    fs.unlinkSync(req.file.path)
+                } else {
+                    const sessObj = {
+                        "_id": new mongoose.Types.ObjectId(),
+                        "created_by": req.session.passport.user,
+                        "data": data._id
+                    }
+                    const newSess = new Sessions(sessObj)
+                    newSess.save((err, user) => {
+                        if (err) {
+                            res.render('sessions', { failed: 'File Uploaded Failed' })
+                        } else {
+                            res.render('sessions', { succes: 'Session Created Succesfully' })
+                        }
+                    })
+                    fs.unlinkSync(req.file.path)
+                }
+            })
+        }
+        fs.unlinkSync(req.file.path)
     } catch (error) {
         fs.unlinkSync(req.file.path)
         res.render('sessions', { failed: 'File Uploaded Failed' })
